+++ src/Storage/PublicSuffixListStorageFactory.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace Pdp\Storage;
 
+use DateInterval;
+use DateTimeInterface;
+
 interface PublicSuffixListStorageFactory
 {
     /**
-     * @param mixed $cacheTtl The cache TTL
+     * @param DateInterval|DateTimeInterface|object|int|string|null $cacheTtl storage TTL object should implement the __toString method
      */
     public function createPublicSuffixListStorage(string $cachePrefix = '', $cacheTtl = null): PublicSuffixListStorage;
 }