+++ src/Storage/TopLevelDomainListStorageFactory.php
@@ -5,10 +5,13 @@ declare(strict_types=1);
 
 namespace Pdp\Storage;
 
+use DateInterval;
+use DateTimeInterface;
+
 interface TopLevelDomainListStorageFactory
 {
     /**
-     * @param mixed $cacheTtl The cache TTL
+     * @param DateInterval|DateTimeInterface|object|int|string|null $cacheTtl storage TTL object should implement the __toString method
      */
     public function createTopLevelDomainListStorage(string $cachePrefix = '', $cacheTtl = null): TopLevelDomainListStorage;
 }